+++ cli/k8s_client/yaml_factory.go
@@ -422,7 +422,7 @@ func GetCSIDeploymentYAML(args *DeploymentYAMLArguments) string {
 
 	csiSnapshotterVersion := "v3.0.3"
 	if args.SnapshotCRDVersion == "v1" {
-		csiSnapshotterVersion = "v6.1.0"
+		csiSnapshotterVersion = "v6.2.1"
 	}
 
 	if args.ImageRegistry == "" {
@@ -584,7 +584,7 @@ spec:
         - name: asup-dir
           mountPath: /asup
       - name: csi-provisioner
-        image: {CSI_SIDECAR_REGISTRY}/csi-provisioner:v3.3.0
+        image: {CSI_SIDECAR_REGISTRY}/csi-provisioner:v3.4.0
         imagePullPolicy: {IMAGE_PULL_POLICY}
         args:
         - "--v={LOG_LEVEL}"
@@ -600,7 +600,7 @@ spec:
         - name: socket-dir
           mountPath: /var/lib/csi/sockets/pluginproxy/
       - name: csi-attacher
-        image: {CSI_SIDECAR_REGISTRY}/csi-attacher:v4.0.0
+        image: {CSI_SIDECAR_REGISTRY}/csi-attacher:v4.1.0
         imagePullPolicy: {IMAGE_PULL_POLICY}
         args:
         - "--v={LOG_LEVEL}"
@@ -614,7 +614,7 @@ spec:
         - name: socket-dir
           mountPath: /var/lib/csi/sockets/pluginproxy/
       - name: csi-resizer
-        image: {CSI_SIDECAR_REGISTRY}/csi-resizer:v1.6.0
+        image: {CSI_SIDECAR_REGISTRY}/csi-resizer:v1.7.0
         imagePullPolicy: {IMAGE_PULL_POLICY}
         args:
         - "--v={LOG_LEVEL}"
@@ -880,7 +880,7 @@ spec:
           mountPath: /certs
           readOnly: true
       - name: driver-registrar
-        image: {CSI_SIDECAR_REGISTRY}/csi-node-driver-registrar:v2.5.1
+        image: {CSI_SIDECAR_REGISTRY}/csi-node-driver-registrar:v2.7.0
         imagePullPolicy: {IMAGE_PULL_POLICY}
         args:
         - "--v={LOG_LEVEL}"
@@ -1059,7 +1059,7 @@ spec:
             cpu: 10m
             memory: 20Mi
       - name: node-driver-registrar
-        image: {CSI_SIDECAR_REGISTRY}/csi-node-driver-registrar:v2.5.1
+        image: {CSI_SIDECAR_REGISTRY}/csi-node-driver-registrar:v2.7.0
         imagePullPolicy: {IMAGE_PULL_POLICY}
         args:
         - --v=2